fix new command

+++ main.py
@@ -18,39 +18,6 @@ help_message = "Список команд для бота:\n" \
                "Канал с демотиваторами из бота: https://t.me/Jacque_Fresco_memes"
 
 
-@dp.message_handler(commands=["dm"])
-async def generate_modern(message: types.Message):
-    try:
-        file_name = f"photos/{message.chat.id}_{time.time()}.jpg"
-        textch = await textcheck(message)
-        attachments = await check(message)
-        if attachments == "message_photo":
-            if textch is not None:
-                await message.photo[-1].download(destination=file_name)
-        elif attachments == "reply_photo":
-            if textch is not None:
-                await message.reply_to_message.photo[-1].download(destination=file_name)
-        elif attachments == "sticker":
-            if textch is not None:
-                await message.reply_to_message.sticker.download(destination=file_name)
-        elif attachments == "anim_sticker":
-            await bot.send_message(chat_id=message.chat.id, text=f"Анимированный стикер не подходит!")
-        else:
-            await bot.send_message(chat_id=message.chat.id, text=f"А где картинка")
-            return
-        if textch is None:
-            return
-        if len(textch["text"]) == 1:
-            media = InputFile(auto(file_name, textch["text"][0]))
-        elif len(textch["text"]) >= 2:
-            media = InputFile(auto(file_name, textch["text"][0], textch["text"][1]))
-        await bot.send_photo(chat_id=message.chat.id, photo=media, reply_to_message_id=message.message_id)
-        os.remove(file_name)
-        await statistics_write(message.chat.id)
-    except Exception as e:
-        await bot.send_message(chat_id=message.chat.id, text=f"ERROR код ошибки \n{e}")
-
-
 @dp.message_handler(commands=["dmc"])
 async def generate_classic(message: types.Message):
     try:
@@ -77,6 +44,39 @@ async def generate_classic(message: types.Message):
             media = InputFile(classic(file_name, textch["text"][0], ''))
         elif len(textch["text"]) >= 2:
             media = InputFile(classic(file_name, textch["text"][0], textch["text"][1]))
+        await bot.send_photo(chat_id=message.chat.id, photo=media, reply_to_message_id=message.message_id)
+        os.remove(file_name)
+        await statistics_write(message.chat.id)
+    except Exception as e:
+        await bot.send_message(chat_id=message.chat.id, text=f"ERROR код ошибки \n{e}")
+
+
+@dp.message_handler(commands=["dm"])
+async def generate_modern(message: types.Message):
+    try:
+        file_name = f"photos/{message.chat.id}_{time.time()}.jpg"
+        textch = await textcheck(message)
+        attachments = await check(message)
+        if attachments == "message_photo":
+            if textch is not None:
+                await message.photo[-1].download(destination=file_name)
+        elif attachments == "reply_photo":
+            if textch is not None:
+                await message.reply_to_message.photo[-1].download(destination=file_name)
+        elif attachments == "sticker":
+            if textch is not None:
+                await message.reply_to_message.sticker.download(destination=file_name)
+        elif attachments == "anim_sticker":
+            await bot.send_message(chat_id=message.chat.id, text=f"Анимированный стикер не подходит!")
+        else:
+            await bot.send_message(chat_id=message.chat.id, text=f"А где картинка")
+            return
+        if textch is None:
+            return
+        if len(textch["text"]) == 1:
+            media = InputFile(auto(file_name, textch["text"][0]))
+        elif len(textch["text"]) >= 2:
+            media = InputFile(auto(file_name, textch["text"][0], textch["text"][1]))
         await bot.send_photo(chat_id=message.chat.id, photo=media, reply_to_message_id=message.message_id)
         os.remove(file_name)
         await statistics_write(message.chat.id)
